requirements.txt
gradio
gradio_modal
geopy
geopandas
pillow
python-dotenv
datasets
huggingface_hub
